requirements.txt
attrs==20.3.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.1
configparser==4.0.2
cryptography==3.2.1
et-xmlfile==1.0.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
google-api-core==1.15.0
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.1.0
google-cloud-storage==1.24.1
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
gunicorn==20.0.4
httplib2==0.15.0
idna==2.8
iniconfig==1.1.1
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
lxml==4.4.2
MarkupSafe==1.1.1
numpy==1.18.1
oauthlib==3.1.0
oci==2.26.0
openpyxl==3.0.3
packaging==20.8
pdfminer.six==20200104
Pillow==7.0.0
pluggy==0.13.1
protobuf==3.11.2
psycopg2-binary==2.8.4
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
pycryptodome==3.9.4
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pytest==6.1.1
pytest-mock==3.4.0
python-dateutil==2.8.1
python-decouple==3.3
python-docx==0.8.10
python-pptx==0.6.18
pytz==2019.3
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sortedcontainers==2.1.0
SQLAlchemy==1.3.16
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.7
Werkzeug==1.0.1
XlsxWriter==1.2.7
